Create method of TODO #1

+++ src/main/java/com/parkit/parkingsystem/service/FareCalculatorService.java
@@ -13,7 +13,7 @@ public class FareCalculatorService {
         long inTime = ticket.getInTime().getTime(); // Time in milliseconds
         long outTime = ticket.getOutTime().getTime(); // Time in milliseconds
 
-        double durationTime = outTime - inTime;
+        double durationTime = (double) outTime - inTime;
         // milliseconds - (/1000) -> seconds - (/60) -> minutes - (/60) -> hours
         double duration = ((durationTime / 1000) / 60) / 60;
